Move Box code into its own file
